+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Rails 3.X and 4.0 no longer supported
 * Ruby 1.9 no longer officially supported (it may still work)
+* Isolated unit tests (via `stub_url`) no longer supported. Use the
+  Spring preloader to keep your app warm instead.
 
 ## Version 1.2.1 ##